back to stol, works now for some reason

+++ assembler.cpp
@@ -594,7 +594,7 @@ int main(int argc, char* argv[]) {
     ofstream file;
     file.open(argv[argc-1], ios::binary);
     for(int i = 0; i < binaryString.size(); i++) {
-        int toWrite = stoi(binaryString[i],nullptr,2);
+        int toWrite = stol(binaryString[i],nullptr,2);
         file.write((char *) &toWrite, sizeof(int));
         // bitset<32> line = bitset<32>binaryString[i];
         // file.write((char *) &line, sizeof(int));
